requirements.txt
streamlit==1.28.0
lyricsgenius==3.0.1
wordcloud==1.9.2
matplotlib==3.7.2
pandas==2.0.3
Pillow==10.0.0
requests==2.31.0